Magzine 样式微调
