+++ themes/magzine/components/ArticleInfo.js
@@ -14,7 +14,7 @@ export default function ArticleInfo(props) {
 
   return (
     <>
-      <div className='flex flex-col gap-y-4 py-4'>
+      <div className='flex flex-col gap-y-4 py-4 px-2 lg:px-0'>
         <div className='flex justify-center mr-2 items-center'>
           {siteConfig('MAGZINE_POST_LIST_CATEGORY') && (
             <CategoryItem category={post?.category} />
